Move some initialize and check hit-object related method into the base change handler class because change the property in the beatmap might affect the change the hit-object. E.g. remove the singer from the beatmap should remove the singer id from the lyric.
