+++ osu.Game.Rulesets.Karaoke.Tests/Editor/ChangeHandlers/BaseChangeHandlerTest.cs
@@ -2,12 +2,15 @@
 // See the LICENCE file in the repository root for full licence text.
 
 using System;
+using System.Collections.Generic;
+using System.Linq;
 using NUnit.Framework;
 using osu.Framework.Allocation;
 using osu.Framework.Graphics;
 using osu.Framework.Testing;
 using osu.Game.Rulesets.Karaoke.Beatmaps;
 using osu.Game.Rulesets.Karaoke.Beatmaps.Metadatas;
+using osu.Game.Rulesets.Objects;
 using osu.Game.Screens.Edit;
 using osu.Game.Tests.Visual;
 
@@ -129,6 +132,50 @@ namespace osu.Game.Rulesets.Karaoke.Tests.Editor.ChangeHandlers
 
                 assert.Invoke(karaokeBeatmap);
             });
+        }
+
+        protected void PrepareHitObject(HitObject hitObject, bool selected = true)
+            => PrepareHitObjects(new[] { hitObject }, selected);
+
+        protected void PrepareHitObjects(IEnumerable<HitObject> selectedHitObjects, bool selected = true)
+        {
+            AddStep("Prepare testing hit objects", () =>
+            {
+                var hitobjects = selectedHitObjects.ToList();
+                var editorBeatmap = Dependencies.Get<EditorBeatmap>();
+
+                editorBeatmap.AddRange(hitobjects);
+
+                if (selected)
+                {
+                    editorBeatmap.SelectedHitObjects.AddRange(hitobjects);
+                }
+            });
+        }
+
+        protected void AssertHitObject<THitObject>(Action<THitObject> assert) where THitObject : HitObject
+        {
+            AssertHitObjects<THitObject>(hitObjects =>
+            {
+                foreach (var hitObject in hitObjects)
+                {
+                    assert(hitObject);
+                }
+            });
+        }
+
+        protected void AssertHitObjects<THitObject>(Action<IEnumerable<THitObject>> assert) where THitObject : HitObject
+        {
+            AddStep("Is result matched", () =>
+            {
+                var editorBeatmap = Dependencies.Get<EditorBeatmap>();
+                assert(editorBeatmap.HitObjects.OfType<THitObject>());
+            });
+
+            // even if there's no property changed in the lyric editor, should still trigger the change handler.
+            // because every change handler call should cause one undo step.
+            // also, technically should not call the change handler if there's no possible to change the properties.
+            AssertTransactionOnlyTriggerOnce();
         }
 
         protected void AssertTransactionOnlyTriggerOnce()
